Finish Async Request
Former-commit-id: a8f6036007e47d6e5cf680cda3468ef3f405226e [formerly 3a73c3bf8f41a4d964189602dee1d5dfa132d3e4] [formerly e479171443281bd281c3b6ae595bd6dfea9a5c3b [formerly 50c2a9c7508118c19003d0e3ccf859241385b7a0]]
Former-commit-id: d126ec0cb56b5d8dd53ef56827041b95cd34ad6b [formerly 0c9c01b9fe5632de052ad4385fb7afae9f6aac0e]
Former-commit-id: 7c179441b5ad03d2182ba516fd6fcf986a5e36d6

+++ controllers/admin.js
@@ -1,15 +1,33 @@
 const Product = require('../models/product')
+const fileHandler = require('../util/file')
+const path = require('path')
+
+const ITEMS_PER_PAGE = 2
 
 exports.getProducts = async (req, res, next) => {
 	try {
+		const count = await Product.find({ userId: req.user.id }).countDocuments()
+		let page = 1
+		if (req.query.page) {
+			page = parseInt(req.query.page)
+		}
 		const products = await Product.find({ userId: req.user.id })
+			.skip((page - 1) * ITEMS_PER_PAGE)
+			.limit(ITEMS_PER_PAGE)
 		res.render('admin/product-list', {
 			products,
-			active: '/admin' + req.url,
-			pageTitle: 'Admin Product List'
+			active: '/admin/product-list',
+			pageTitle: 'Admin Product List',
+			path: '/admin/product-list',
+			currentPage: page,
+			nextPage: page + 1,
+			prevPage: page - 1,
+			lastPage: Math.ceil(count / ITEMS_PER_PAGE)
 		})
 	} catch (err) {
-		console.log(err)
+		const error = new Error(err)
+		error.httpStatusCode = 500
+		return next(error)
 	}
 }
 
@@ -22,18 +40,28 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = async (req, res, next) => {
 	try {
+		const image = req.file
+		if (!image) {
+			return res.render('admin/add-product', {
+				active: '/admin' + req.url,
+				pageTitle: 'Add Product',
+				errMsg: 'Invalid Image File'
+			})
+		}
 		const product = new Product({
 			id: Date.now().toString(),
 			title: req.body.title,
 			price: req.body.price,
-			imgUrl: req.body.imgUrl,
+			imgUrl: '/' + image.path,
 			description: req.body.description,
 			userId: req.user.id
 		})
 		await product.save()
 		res.redirect('/admin/add-product')
 	} catch (err) {
-		console.log(err)
+		const error = new Error(err)
+		error.httpStatusCode = 500
+		return next(error)
 	}
 }
 
@@ -49,7 +77,9 @@ exports.getEditProduct = async (req, res, next) => {
 			product
 		})
 	} catch (err) {
-		console.log(err)
+		const error = new Error(err)
+		error.httpStatusCode = 500
+		return next(error)
 	}
 }
 
@@ -58,28 +88,54 @@ exports.postEditProduct = async (req, res, next) => {
 		const update = {
 			title: req.body.title,
 			price: req.body.price,
-			imgUrl: req.body.imgUrl,
 			description: req.body.description
+		}
+		const image = req.file
+		if (image) {
+			update.imgUrl = '/' + image.path
+			const product = await Product.findOne({
+				id: req.body.id,
+				userId: req.user.id
+			})
+			const filePath = path.join(__dirname, '..', product.imgUrl)
+			fileHandler.deleteFile(filePath)
 		}
 		await Product.updateOne(
 			{ id: req.body.id, userId: req.user.id },
 			update,
 			err => {
-				if (err) console.log(err)
+				if (err) {
+					const error = new Error(err)
+					error.httpStatusCode = 500
+					return next(error)
+				}
 			}
 		)
 		res.redirect('/')
 	} catch (err) {
-		console.log(err)
+		const error = new Error(err)
+		error.httpStatusCode = 500
+		return next(error)
 	}
 }
 
-exports.getDeleteProduct = async (req, res, next) => {
+exports.deleteProduct = async (req, res, next) => {
+	const product = await Product.findOne({
+		id: req.params.productId,
+		userId: req.user.id
+	})
+	if (!product) {
+		return res.status(500).json({ message: 'Failed' })
+	}
+	const filePath = path.join(__dirname, '..', product.imgUrl)
+	fileHandler.deleteFile(filePath)
 	await Product.deleteOne(
 		{ id: req.params.productId, userId: req.user.id },
 		err => {
-			if (err) console.log(err)
+			if (err) {
+				return res.status(500).json({ message: 'Failed' })
+			}
 		}
 	)
-	res.redirect('/admin/product-list')
+	res.status(200).json({ message: 'Success' })
 }
